+++ src/main/java/io/github/mzmine/modules/visualization/rawdataoverviewims/IMSRawDataOverviewPane.java
@@ -33,15 +33,15 @@ import io.github.mzmine.gui.preferences.UnitFormat;
 import io.github.mzmine.main.MZmineCore;
 import io.github.mzmine.modules.visualization.chromatogram.TICDataSet;
 import io.github.mzmine.modules.visualization.chromatogram.TICPlot;
-import io.github.mzmine.modules.visualization.rawdataoverviewims.Threads.BuildMultipleMobilogramRanges;
-import io.github.mzmine.modules.visualization.rawdataoverviewims.Threads.BuildMultipleTICRanges;
-import io.github.mzmine.modules.visualization.rawdataoverviewims.Threads.BuildSelectedRanges;
 import io.github.mzmine.modules.visualization.rawdataoverviewims.providers.CachedFrame;
 import io.github.mzmine.modules.visualization.rawdataoverviewims.providers.FrameHeatmapProvider;
 import io.github.mzmine.modules.visualization.rawdataoverviewims.providers.FrameSummedMobilogramProvider;
 import io.github.mzmine.modules.visualization.rawdataoverviewims.providers.FrameSummedSpectrumProvider;
 import io.github.mzmine.modules.visualization.rawdataoverviewims.providers.IonTraceProvider;
 import io.github.mzmine.modules.visualization.rawdataoverviewims.providers.SingleSpectrumProvider;
+import io.github.mzmine.modules.visualization.rawdataoverviewims.threads.BuildMultipleMobilogramRanges;
+import io.github.mzmine.modules.visualization.rawdataoverviewims.threads.BuildMultipleTICRanges;
+import io.github.mzmine.modules.visualization.rawdataoverviewims.threads.BuildSelectedRanges;
 import io.github.mzmine.parameters.parametertypes.selectors.ScanSelection;
 import io.github.mzmine.parameters.parametertypes.tolerances.MZTolerance;
 import io.github.mzmine.project.impl.StorableFrame;
